+++ frontend/reseau-social-groupomania/src/views/Profile.vue
@@ -162,8 +162,13 @@ export default {
             this.description = data.data[0].description;
             })
             .catch(error => {
-                this.errorMessage = error;
-                console.error("There was an error!", error);
+                if(error == 'Unauthorized'){
+                    alert('Votre session a expiré vous allez être redirigé vers la page de connexion')
+                    sessionStorage.clear();
+                    window.location.href = '/';
+                }else{
+                    console.error('There was an error!', error);
+                }
             });
         },
         modify: function(){
@@ -179,7 +184,6 @@ export default {
                     password: this.password,
                     description: this.description,
                 }),
-                
             })
             .then(async res => {
                 const data = await res.json();
@@ -191,11 +195,13 @@ export default {
                 this.displayProfile();
             })
             .catch(error => {
-                this.errorMessage = error;
-                console.error("There was an error!", error);
-                alert('Votre session a expirée vous allez être redirigé vers la page de connexion')
-                sessionStorage.clear();
-                window.location.href = '/';
+                if(error == 'Unauthorized'){
+                    alert('Votre session a expiré vous allez être redirigé vers la page de connexion')
+                    sessionStorage.clear();
+                    window.location.href = '/';
+                }else{
+                    console.error('There was an error!', error);
+                }
             });
         },
         onFileChange() {
@@ -221,11 +227,13 @@ export default {
             location.reload();
             })
             .catch(error => {
-                this.errorMessage = error;
-                console.error("There was an error!", error);
-                alert('Votre session a expirée vous allez être redirigé vers la page de connexion')
-                sessionStorage.clear();
-                window.location.href = '/';
+                if(error == 'Unauthorized'){
+                    alert('Votre session a expiré vous allez être redirigé vers la page de connexion')
+                    sessionStorage.clear();
+                    window.location.href = '/';
+                }else{
+                    console.error('There was an error!', error);
+                }
             });
         }, 
         deletePicture: function(element){
@@ -248,11 +256,13 @@ export default {
             this.displayProfile();
             })
             .catch(error => {
-                this.errorMessage = error;
-                console.error("There was an error!", error);
-                alert('Votre session a expirée vous allez être redirigé vers la page de connexion')
-                sessionStorage.clear();
-                window.location.href = '/';
+                if(error == 'Unauthorized'){
+                    alert('Votre session a expiré vous allez être redirigé vers la page de connexion')
+                    sessionStorage.clear();
+                    window.location.href = '/';
+                }else{
+                    console.error('There was an error!', error);
+                }
             });
             }
         },
